HDDS-9488. Merge TestOzoneFileSystemWithLinks into TestOzoneFileSystem
